Move the win32 implementation to platform/windows
