+++ src/api/mod.rs
@@ -3,6 +3,5 @@ pub mod cocoa;
 pub mod dlopen;
 pub mod emscripten;
 pub mod wayland;
-pub mod win32;
 pub mod x11;
 pub mod ios;
