+++ src/platform/windows/callback.rs
